Drap&drop selection unification (aurusov/raox#366)
